+++ src/marsyas/marsystems/RadioDrumInput.cpp
@@ -25,6 +25,7 @@
 #include "RtMidi.h"
 #endif
 
+#include <cstddef>
 
 using namespace std;
 using namespace Marsyas;
@@ -143,7 +144,7 @@ void RadioDrumInput::mycallback(double deltatime, std::vector< unsigned char > *
 {
   // FIXME Unused parameter
   (void) deltatime;
-  int nBytes = 0;
+  size_t nBytes = 0;
   nBytes = message->size();
 
   RadioDrumInput* mythis = (RadioDrumInput*) userData;
